requirements.txt
python-telegram-bot>=20.0
anthropic>=0.18.0
aiosqlite>=0.19.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
pytz>=2024.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
